Move the "query" part from Joining\AbstractQuery to a separate trait so we can reuse it
