+++ src/Search/Query/Joining/AbstractQuery.php
@@ -2,6 +2,7 @@
 
 use Nord\Lumen\Elasticsearch\Exceptions\InvalidArgument;
 use Nord\Lumen\Elasticsearch\Search\Query\QueryDSL;
+use Nord\Lumen\Elasticsearch\Search\Query\Traits\HasQuery;
 
 /**
  * Performing full SQL-style joins in a distributed system like Elasticsearch is prohibitively expensive.
@@ -20,18 +21,14 @@ use Nord\Lumen\Elasticsearch\Search\Query\QueryDSL;
  */
 abstract class AbstractQuery extends QueryDSL
 {
-
+    use HasQuery;
+    
     const SCORE_MODE_AVG   = 'avg';
     const SCORE_MODE_SUM   = 'sum';
     const SCORE_MODE_MIN   = 'min';
     const SCORE_MODE_MAX   = 'max';
     const SCORE_MODE_SCORE = 'score';
     const SCORE_MODE_NONE  = 'none';
-
-    /**
-     * @var QueryDSL;
-     */
-    protected $query;
 
     /**
      * @var string
@@ -42,26 +39,6 @@ abstract class AbstractQuery extends QueryDSL
      * @return array
      */
     abstract protected function getValidScoreModes();
-
-    /**
-     * @return QueryDSL
-     */
-    public function getQuery()
-    {
-        return $this->query;
-    }
-
-    /**
-     * @param QueryDSL $query
-     *
-     * @return $this
-     */
-    public function setQuery(QueryDSL $query)
-    {
-        $this->query = $query;
-
-        return $this;
-    }
 
     /**
      * @return string
